refactor: Refactor packages and directories
- Add security package
