+++ src/main/java/nextstep/app/configuration/ControllerExceptionHandler.java
@@ -1,6 +1,6 @@
 package nextstep.app.configuration;
 
-import nextstep.app.exception.AuthenticationException;
+import nextstep.security.exception.AuthenticationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
