+++ src/main/java/nextstep/security/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.app.exception;
+package nextstep.security.exception;
 
 public class AuthenticationException extends RuntimeException {
 }
